Modules/requirements.txt
scikit-learn
pytorch-lightning
numpy
matplotlib
seaborn
torchaudio
speechbrain


